backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "async-timeout==5.0.1",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "diff-match-patch==20241021",
    "django==5.2.4",
    "django-ckeditor==6.7.3",
    "django-cors-headers==4.7.0",
    "django-extensions==4.1",
    "django-filter==25.1",
    "django-import-export==4.3.8",
    "django-js-asset==3.1.2",
    "django-otp==1.6.1",
    "django-redis==6.0.0",
    "django-seed==0.3.1",
    "django-unfold==0.62.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.3.0",
    "faker==37.5.3",
    "idna==3.10",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.10.1",
    "python-decouple==3.8",
    "qrcode==8.2",
    "redis==6.2.0",
    "requests==2.32.4",
    "sqlparse==0.5.3",
    "tablib==3.8.0",
    "toposort==1.10",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
]
